Renaming to adhere to convention  - Modify CMakeList to compile using ROS2  - TODO: fix FMT import
