+++ system/autoware_error_monitor/include/autoware_error_monitor/autoware_error_monitor_core.h
