--- system/autoware_error_monitor/src/autoware_error_monitor_node/main.cpp
@@ -1,30 +1,0 @@
-/*
- * Copyright 2020 Tier IV, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include <ros/ros.h>
-
-#include <autoware_error_monitor/autoware_error_monitor_node.h>
-
-int main(int argc, char * argv[])
-{
-  ros::init(argc, argv, "autoware_error_monitor");
-
-  AutowareErrorMonitorNode node;
-
-  ros::spin();
-
-  return 0;
-}
